Refactor Ruby classes into separate files
