+++ _data/meetups.yml
@@ -1429,7 +1429,7 @@
   end_time: 20:00:00 EDT
   url: https://lu.ma/af2n3jwa
 
-- name: "Austin.rb - Visualizing Problems: From Abstract to Concrete"
+- name: 'Austin.rb - Visualizing Problems: From Abstract to Concrete'
   location: Austin, TX
   date: 2024-09-10
   start_time: 18:30:00 CDT
