+++ src/static/conference.rb
@@ -1,0 +1,4 @@
+require_relative "../static"
+
+class Conference < FrozenRecord::Base
+end
